Backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
requests
python-dotenv
Flask-Cors
pandas
